improved the generated pdf document

+++ controllers/admin.js
@@ -1,6 +1,7 @@
 const Product = require('../models/product');
 const getFlashErrorMessages = require('../util/getFlashErrorMessage');
 const { validationResult } = require('express-validator');
+const fileOperations = require('../util/fileOperations');
 
 exports.getProducts = (req, res, next) => {
 	Product.find({ userId: req.user._id })
@@ -33,7 +34,7 @@ exports.postAddProduct = (req, res, next) => {
 	const uploadedImageFile = req.file;
 	const price = req.body.price;
 	const description = req.body.description;
-	
+
 	// validation errors on user input
 	const errors = validationResult(req);
 	if (!errors.isEmpty()) {
@@ -55,7 +56,7 @@ exports.postAddProduct = (req, res, next) => {
 	const product = new Product({
 		title: title,
 		price: price,
-		imagePath: '/' + uploadedImageFile.path,
+		imagePath: '\\' + uploadedImageFile.path,
 		description: description,
 		userId: req.user, // in mongoose it is possible to reference the whole user object, mongoose takes the id from this object
 	});
@@ -110,7 +111,6 @@ exports.postEditProduct = (req, res, next) => {
 	const updatedUploadedImageFile = req.file;
 	const updatedDescription = req.body.description;
 
-	
 	const errors = validationResult(req);
 	if (!errors.isEmpty()) {
 		return res.status(422).render('admin/edit-product', {
@@ -137,7 +137,10 @@ exports.postEditProduct = (req, res, next) => {
 			product.title = updatedTitle;
 			product.price = updatedPrice;
 			if (updatedUploadedImageFile) {
-				product.imagePath = '/' + updatedUploadedImageFile.path;
+				//delete old image
+				fileOperations.deleteFile(product.imagePath);
+
+				product.imagePath = '\\' + updatedUploadedImageFile.path;
 			}
 			product.description = updatedDescription;
 			return product.save();
@@ -156,12 +159,24 @@ exports.postEditProduct = (req, res, next) => {
 exports.postDeleteProduct = (req, res, next) => {
 	const productId = req.body.productId;
 
-	Product.deleteOne({ _id: productId, userId: req.user._id })
+	Product.findById(productId)
 		.then((product) => {
+			if (!product) {
+				return next(new Error('Product not found!'));
+			}
+			//delete image from server
+			fileOperations.deleteFile(product.imagePath);
+			//delete product from db
+			return Product.deleteOne({ _id: productId, userId: req.user._id });
+		})
+		.then((result) => {
+			console.log(`PRODUCT WITH _id:${productId} SUCCESSFULLY DELETED`);
 			res.redirect('/admin/products');
 		})
 		.catch((err) => {
-			const error = new Error();
+			console.log(`ERROR ON PRODUCT _id:${productId} DELETION`);
+			console.log(err);
+			const error = new Error(err);
 			error.httpStatusCode = 500;
 			return next(error);
 		});
